Fixed PID controller and re-tuned
Fixed issues with the PID controller using the micros function. Issues was the max value that time divisor could hold. Changed data type to double. Changed some parameters around.
Re-did how the swing works. Now it uses a counter that cycles on every sample instead of using multiple micros functions that need to be changed.

+++ Controller_Code/PID.h
@@ -17,7 +17,7 @@ class Pid
 {
 	private:
 		int TimeStep_;
-		int TimeDivisor_;
+		double TimeDivisor_;
 		double Kp_;
 		double Ki_;
 		double Kd_;
